refactor(semantic): split reference collection from store
